[#21] apply ErrorPage component
